@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,12 +27,18 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  firebase_auth: ^1.1.1
+  firebase_core: ^1.0.4
+  cloud_firestore: ^1.0.6
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  amplify_auth_cognito: ^0.1.2
-  amplify_flutter: ^0.1.2
+  cloud_functions: ^1.0.4
+  firebase_admin: ^0.1.4
+
+dependency_overrides:
+  http_parser: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
